refactor: Ajustando pasta do CertificationComponent e suas referências
