+++ src/app/pages/resume/resume.component.ts
@@ -3,7 +3,7 @@ import { SharedModule } from '../../shared/shared.module';
 import { EducationComponent } from '../../modules/education/components/education/education.component';
 import { ExperienceComponent } from "../../modules/experience/components/experience/experience.component";
 import { SkillComponent } from "../../modules/skill/components/skill/skill.component";
-import { CertificationComponent } from "../../modules/certification/certification/certification.component";
+import { CertificationComponent } from "../../modules/certification/components/certification/certification.component";
 
 @Component({
   selector: 'app-resume',
